Extracted Tool Initialization from Editor
